sourcepp: adopt LERep from vtfpp
